+++ src/main/java/kitchenpos/order/domain/OrderLineItem.java
@@ -1,0 +1,54 @@
+package kitchenpos.order.domain;
+
+public class OrderLineItem {
+    private Long seq;
+    private Long orderId;
+    private Long menuId;
+    private long quantity;
+
+    public OrderLineItem() {
+
+    }
+
+    private OrderLineItem(Long seq, Long orderId, Long menuId, long quantity) {
+        this.seq = seq;
+        this.orderId = orderId;
+        this.menuId = menuId;
+        this.quantity = quantity;
+    }
+    public static OrderLineItem of(Long seq, Long orderId, Long menuId, long quantity) {
+        return new OrderLineItem(seq, orderId, menuId, quantity);
+    }
+
+    public Long getSeq() {
+        return seq;
+    }
+
+    public void setSeq(final Long seq) {
+        this.seq = seq;
+    }
+
+    public Long getOrderId() {
+        return orderId;
+    }
+
+    public void setOrderId(final Long orderId) {
+        this.orderId = orderId;
+    }
+
+    public Long getMenuId() {
+        return menuId;
+    }
+
+    public void setMenuId(final Long menuId) {
+        this.menuId = menuId;
+    }
+
+    public long getQuantity() {
+        return quantity;
+    }
+
+    public void setQuantity(final long quantity) {
+        this.quantity = quantity;
+    }
+}
